Refactor(bbs) : cop.bbs 전체 패키지 구조를 MVC 패턴에 기반하여 역할별 패키지 재구성   - mapper 파일 내 service 객체 참조 패키지 경로 수정
