+++ src/main/java/egovframework/let/cop/bbs/repository/BBSAttributeManageDAO.java
@@ -1,12 +1,12 @@
-package egovframework.let.cop.bbs.service.impl;
+package egovframework.let.cop.bbs.repository;
 import java.util.List;
-
-import egovframework.let.cop.bbs.service.BoardMaster;
-import egovframework.let.cop.bbs.service.BoardMasterVO;
 
 import org.egovframe.rte.psl.dataaccess.EgovAbstractMapper;
 
 import org.springframework.stereotype.Repository;
+
+import egovframework.let.cop.bbs.domain.model.BoardMaster;
+import egovframework.let.cop.bbs.domain.model.BoardMasterVO;
 
 /**
  * 게시판 속성정보 관리를 위한 데이터 접근 클래스
