+++ src/main/java/egovframework/let/cop/bbs/repository/BBSManageDAO.java
@@ -1,13 +1,13 @@
-package egovframework.let.cop.bbs.service.impl;
+package egovframework.let.cop.bbs.repository;
 import java.util.Iterator;
 import java.util.List;
-
-import egovframework.let.cop.bbs.service.Board;
-import egovframework.let.cop.bbs.service.BoardVO;
 
 import org.egovframe.rte.psl.dataaccess.EgovAbstractMapper;
 
 import org.springframework.stereotype.Repository;
+
+import egovframework.let.cop.bbs.domain.model.Board;
+import egovframework.let.cop.bbs.domain.model.BoardVO;
 
 /**
  * 게시물 관리를 위한 데이터 접근 클래스
