+++ src/main/java/egovframework/let/cop/bbs/service/EgovBBSManageService.java
@@ -2,6 +2,9 @@ package egovframework.let.cop.bbs.service;
 
 import java.util.Map;
 
+import egovframework.let.cop.bbs.domain.model.Board;
+import egovframework.let.cop.bbs.domain.model.BoardVO;
+
 /**
  * 게시물 관리를 위한 서비스 인터페이스  클래스
  * @author 공통 서비스 개발팀 이삼섭
