+++ src/main/java/egovframework/let/cop/bbs/service/impl/EgovBBSManageServiceImpl.java
@@ -12,8 +12,9 @@ import org.springframework.stereotype.Service;
 
 import egovframework.com.cmm.service.EgovFileMngService;
 import egovframework.com.cmm.service.FileVO;
-import egovframework.let.cop.bbs.service.Board;
-import egovframework.let.cop.bbs.service.BoardVO;
+import egovframework.let.cop.bbs.domain.model.Board;
+import egovframework.let.cop.bbs.domain.model.BoardVO;
+import egovframework.let.cop.bbs.repository.BBSManageDAO;
 import egovframework.let.cop.bbs.service.EgovBBSManageService;
 import egovframework.let.utl.fcc.service.EgovDateUtil;
 
@@ -52,7 +53,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 게시물 한 건을 삭제 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#deleteBoardArticle(egovframework.let.cop.bbs.brd.service.Board)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#deleteBoardArticle(egovframework.let.cop.bbs.domain.model.brd.service.Board)
 	 */
 	@Override
 	public void deleteBoardArticle(Board board) throws Exception {
@@ -72,7 +73,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 게시판에 게시물 또는 답변 게시물을 등록 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#insertBoardArticle(egovframework.let.cop.bbs.brd.service.Board)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#insertBoardArticle(egovframework.let.cop.bbs.domain.model.brd.service.Board)
 	 */
 	@Override
 	public void insertBoardArticle(Board board) throws Exception {
@@ -99,7 +100,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 게시물 대하여 상세 내용을 조회 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectBoardArticle(egovframework.let.cop.bbs.brd.service.BoardVO)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectBoardArticle(egovframework.let.cop.bbs.domain.model.brd.service.BoardVO)
 	 */
 	@Override
 	public BoardVO selectBoardArticle(BoardVO boardVO) throws Exception {
@@ -116,7 +117,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 조건에 맞는 게시물 목록을 조회 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectBoardArticles(egovframework.let.cop.bbs.brd.service.BoardVO)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectBoardArticles(egovframework.let.cop.bbs.domain.model.brd.service.BoardVO)
 	 */
 	@Override
 	public Map<String, Object> selectBoardArticles(BoardVO boardVO, String attrbFlag) throws Exception {
@@ -157,7 +158,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 게시물 한 건의 내용을 수정 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#updateBoardArticle(egovframework.let.cop.bbs.brd.service.Board)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#updateBoardArticle(egovframework.let.cop.bbs.domain.model.brd.service.Board)
 	 */
 	@Override
 	public void updateBoardArticle(Board board) throws Exception {
@@ -167,7 +168,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 방명록 내용을 삭제 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#deleteGuestList(egovframework.let.cop.bbs.brd.service.BoardVO)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#deleteGuestList(egovframework.let.cop.bbs.domain.model.brd.service.BoardVO)
 	 */
 	@Override
 	public void deleteGuestList(BoardVO boardVO) throws Exception {
@@ -177,7 +178,7 @@ public class EgovBBSManageServiceImpl extends EgovAbstractServiceImpl implements
 	/**
 	 * 방명록에 대한 목록을 조회 한다.
 	 *
-	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectGuestList(egovframework.let.cop.bbs.brd.service.BoardVO)
+	 * @see egovframework.let.cop.bbs.brd.service.EgovBBSManageService#selectGuestList(egovframework.let.cop.bbs.domain.model.brd.service.BoardVO)
 	 */
 	@Override
 	public Map<String, Object> selectGuestList(BoardVO boardVO) throws Exception {
