+++ src/main/resources/egovframework/mapper/let/cop/bbs/EgovBoard_SQL_oracle.xml
@@ -4,7 +4,7 @@
 <mapper namespace="BBSManageDAO">
 
 
-	<resultMap id="boardList" type="egovframework.let.cop.bbs.service.BoardVO">
+	<resultMap id="boardList" type="egovframework.let.cop.bbs.domain.model.BoardVO">
 		<result property="bbsId" column="BBS_ID"/>
 		<result property="nttId" column="NTT_ID"/>
 		<result property="nttSj" column="NTT_SJ"/>
@@ -21,7 +21,7 @@
 		<result property="ntceEndde" column="NTCE_ENDDE"/>
 	</resultMap>
 
-	<resultMap id="boardDetail" type="egovframework.let.cop.bbs.service.BoardVO">
+	<resultMap id="boardDetail" type="egovframework.let.cop.bbs.domain.model.BoardVO">
 		<result property="bbsId" column="BBS_ID"/>
 		<result property="nttId" column="NTT_ID"/>
 		<result property="nttSj" column="NTT_SJ"/>	
@@ -50,7 +50,7 @@
 		<result property="bbsNm" column="BBS_NM"/>
 	</resultMap>
 
-	<resultMap id="sortList" type="egovframework.let.cop.bbs.service.BoardVO">
+	<resultMap id="sortList" type="egovframework.let.cop.bbs.domain.model.BoardVO">
 		<result property="bbsId" column="BBS_ID"/>
 		<result property="nttId" column="NTT_ID"/>
 		<result property="sortOrdr" column="SORT_ORDR"/>
@@ -59,7 +59,7 @@
 		<result property="replyLc" column="ANSWER_LC"/>
 	</resultMap>
 
-	<resultMap id="guestList" type="egovframework.let.cop.bbs.service.BoardVO">
+	<resultMap id="guestList" type="egovframework.let.cop.bbs.domain.model.BoardVO">
 		<result property="bbsId" column="BBS_ID"/>
 		<result property="nttId" column="NTT_ID"/>
 		<result property="nttSj" column="NTT_SJ"/>
@@ -78,7 +78,7 @@
  		
  	</select>
  	
-	<insert id="insertBoardArticle" parameterType="egovframework.let.cop.bbs.service.Board">
+	<insert id="insertBoardArticle" parameterType="egovframework.let.cop.bbs.domain.model.Board">
 		<selectKey keyProperty="nttNo" resultType="java.lang.Long" order="BEFORE">
 			SELECT NVL(MAX(SORT_ORDR),0)+1 AS NTT_NO  FROM LETTNBBS
 			WHERE BBS_ID = #{bbsId}
@@ -101,7 +101,7 @@
 		
 	</insert>
 	
-	<insert id="replyBoardArticle" parameterType="egovframework.let.cop.bbs.service.Board">
+	<insert id="replyBoardArticle" parameterType="egovframework.let.cop.bbs.domain.model.Board">
 		<selectKey keyProperty="nttNo" resultType="java.lang.Long" order="BEFORE">
 			SELECT NVL(MAX(NTT_NO),0)+1 AS NTT_NO  FROM LETTNBBS
 			WHERE BBS_ID = #{bbsId} AND SORT_ORDR = #{sortOrdr}
@@ -124,7 +124,7 @@
 		
 	</insert>	
 	
-	<select id="selectNoticeListForSort" parameterType="egovframework.let.cop.bbs.service.Board" resultMap="sortList">
+	<select id="selectNoticeListForSort" parameterType="egovframework.let.cop.bbs.domain.model.Board" resultMap="sortList">
 					
 			SELECT
 				a.BBS_ID, a.NTT_ID, a.SORT_ORDR, a.PARNTSCTT_NO, a.NTT_NO, a.ANSWER_LC
@@ -139,7 +139,7 @@
 	</select>
 
 			
-	<select id="selectBoardArticleList" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultMap="boardList">
+	<select id="selectBoardArticleList" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultMap="boardList">
 		
 		SELECT * FROM ( SELECT rownum rn, TB.* FROM (
 			SELECT
@@ -172,7 +172,7 @@
 						
 	</select>	
 	
-	<select id="selectBoardArticleListCnt" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultType="java.lang.Integer">
+	<select id="selectBoardArticleListCnt" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultType="java.lang.Integer">
 		
 			SELECT
 				COUNT(a.NTT_ID)
@@ -197,7 +197,7 @@
 			</if>		
 	</select>	
  
-	<select id="selectBoardArticle" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultMap="boardDetail">
+	<select id="selectBoardArticle" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultMap="boardDetail">
 		
 			SELECT
 				a.NTT_SJ, a.NTCR_ID, a.NTCR_NM, a.NTT_NO, a.NTT_CN,
@@ -225,7 +225,7 @@
 						
 	</select> 
  
- 	<update id="updateBoardArticle" parameterType="egovframework.let.cop.bbs.service.Board">
+ 	<update id="updateBoardArticle" parameterType="egovframework.let.cop.bbs.domain.model.Board">
  		
 			UPDATE LETTNBBS SET 
 				NTT_SJ = #{nttSj},
@@ -243,7 +243,7 @@
  		
  	</update>
 
- 	<update id="deleteBoardArticle" parameterType="egovframework.let.cop.bbs.service.Board">
+ 	<update id="deleteBoardArticle" parameterType="egovframework.let.cop.bbs.domain.model.Board">
  		
 			UPDATE LETTNBBS SET 
 				NTT_SJ = #{nttSj},
@@ -255,7 +255,7 @@
  		
  	</update>
 
- 	<select id="selectMaxInqireCo" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultType="java.lang.Integer">
+ 	<select id="selectMaxInqireCo" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultType="java.lang.Integer">
  		
 			SELECT NVL(MAX(RDCNT),0)+1 AS RDCNT FROM LETTNBBS
 			WHERE BBS_ID = #{bbsId}
@@ -263,7 +263,7 @@
  		
  	</select>
  	
- 	<select id="selectNoticeItemForSort" parameterType="egovframework.let.cop.bbs.service.Board" resultType="java.lang.Long">
+ 	<select id="selectNoticeItemForSort" parameterType="egovframework.let.cop.bbs.domain.model.Board" resultType="java.lang.Long">
  		
 			SELECT
 				NVL(MAX(NTT_NO),0)+1 AS NTT_NO
@@ -278,7 +278,7 @@
  		
  	</select> 	
 
- 	<update id="updateInqireCo" parameterType="egovframework.let.cop.bbs.service.BoardVO">
+ 	<update id="updateInqireCo" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO">
  		
 			UPDATE LETTNBBS SET 
 				RDCNT = #{inqireCo},
@@ -289,7 +289,7 @@
  		
  	</update>
 
-	<update id="updateSortOrder" parameterType="egovframework.let.cop.bbs.service.BoardVO">
+	<update id="updateSortOrder" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO">
  		
 			UPDATE LETTNBBS SET 
 				NTT_NO = #{nttNo}
@@ -297,7 +297,7 @@
  			
 	</update>
 
-	<select id="selectGuestList" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultMap="guestList">
+	<select id="selectGuestList" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultMap="guestList">
 		
 		SELECT * FROM ( SELECT rownum rn, TB.* FROM (
 			SELECT
@@ -318,7 +318,7 @@
 						
 	</select>	
 	
-	<select id="selectGuestListCnt" parameterType="egovframework.let.cop.bbs.service.BoardVO" resultType="java.lang.Integer">
+	<select id="selectGuestListCnt" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO" resultType="java.lang.Integer">
 		
 			SELECT
 				COUNT(NTT_ID)
@@ -331,7 +331,7 @@
 					
 	</select>	
 
-	<update id="deleteGuestList" parameterType="egovframework.let.cop.bbs.service.BoardVO">
+	<update id="deleteGuestList" parameterType="egovframework.let.cop.bbs.domain.model.BoardVO">
 		
 			UPDATE LETTNBBS
 			SET USE_AT = 'N'
@@ -340,7 +340,7 @@
 					
 	</update>	
 	
-	<select id="getPasswordInf" parameterType="egovframework.let.cop.bbs.service.Board" resultType="java.lang.String">
+	<select id="getPasswordInf" parameterType="egovframework.let.cop.bbs.domain.model.Board" resultType="java.lang.String">
 		
 			SELECT
 				PASSWORD
@@ -351,14 +351,14 @@
 		
 	</select>
 
-	<select id="getParentNttNo" parameterType="egovframework.let.cop.bbs.service.Board" resultType="java.lang.Long">
+	<select id="getParentNttNo" parameterType="egovframework.let.cop.bbs.domain.model.Board" resultType="java.lang.Long">
 		
 			SELECT NTT_NO FROM LETTNBBS
 			WHERE BBS_ID = #{bbsId} AND NTT_ID = #{parnts}			
 		
 	</select>
 	
- 	<update id="updateOtherNttNo" parameterType="egovframework.let.cop.bbs.service.Board">
+ 	<update id="updateOtherNttNo" parameterType="egovframework.let.cop.bbs.domain.model.Board">
  		
 			UPDATE LETTNBBS SET 
 				NTT_NO = NTT_NO + 1
@@ -367,7 +367,7 @@
  		
  	</update>
  	
- 	 	<update id="updateNttNo" parameterType="egovframework.let.cop.bbs.service.Board">
+ 	 	<update id="updateNttNo" parameterType="egovframework.let.cop.bbs.domain.model.Board">
  		
 			UPDATE LETTNBBS SET 
 				NTT_NO = #{nttNo}
