+++ src/test/java/egovframework/let/cop/bbs/service/impl/BBSAttributeManageDAOTestInsertBBSMasterInfTest.java
@@ -10,7 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import egovframework.let.cop.bbs.service.BoardMaster;
+import egovframework.let.cop.bbs.domain.model.BoardMaster;
+import egovframework.let.cop.bbs.repository.BBSAttributeManageDAO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
